+++ NEWS.md
@@ -11,6 +11,7 @@ etc falsely returned last season after March 15th early in the month. (1.3.2.03)
 - Clarified description of the `"special"` variable in the play-by-play data dictionary. (1.3.2.04) (#189)
 - `nflreadr::nflverse_game_id()` accepts old team abbreviations and outputs useful errors. (1.3.2.05) (#191)
 - Added `dictionary_player_stats_def` (thank you @mpcen!) (1.3.2.06) (#192)
+- Added `.for_cran()` to limit parallelization in CRAN examples and tests (1.3.2.07)
 
 # nflreadr 1.3.2